Move ClasspathInferer to base-services
